Integrate reports functionality into ProgressPage, remove separate ReportsPage
